Proyecto terminado al 100%

+++ ProyectoMovilesJose/src/main/java/com/empresamoviles/mobiles/dto/MobileDetailDTO.java
@@ -1,37 +1,36 @@
 package com.empresamoviles.mobiles.dto;
 
-import java.math.BigDecimal;
+import lombok.*;
 import java.time.LocalDate;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
 
+/**
+ * DTO con el detalle completo de un móvil.
+ * Incluye todos los campos de la entidad excepto consultationCount,
+ * que es un dato interno y no debe exponerse al cliente.
+ */
 @Data
 @Builder
 @NoArgsConstructor
 @AllArgsConstructor
 public class MobileDetailDTO {
 
-	 private Long id;
-	    private String brand;
-	    private String model;
-	    private String processorType;
-	    private Integer processorCores;
-	    private Double processorMaxSpeedGhz;
-	    private Integer storageGb;
-	    private Double screenSizeInches;
-	    private String screenTechnology;
-	    private Integer ramGb;
-	    private Double heightCm;
-	    private Double widthCm;
-	    private Double thicknessCm;
-	    private Integer weightGrams;
-	    private Integer cameraMp;
-	    private Integer batteryMah;
-	    private Boolean nfc;
-	    private BigDecimal currentPrice;
-	    private LocalDate releaseDate;
-    
-
+    private Long id;
+    private String brand;
+    private String model;
+    private String processorType;
+    private Integer processorCores;
+    private Double processorMaxSpeedGhz;
+    private Integer storageGb;
+    private Double screenSizeInches;
+    private String screenTechnology;
+    private Integer ramGb;
+    private Double heightCm;
+    private Double widthCm;
+    private Double thicknessCm;
+    private Integer weightGrams;
+    private Integer cameraMp;
+    private Integer batteryMah;
+    private Boolean nfc;
+    private Double currentPrice;
+    private LocalDate releaseDate;
 }
